eliminated need for rust api globals and channel busy wait
