+++ core/Cargo-rust.toml
@@ -37,4 +37,3 @@ tokio-stdin-stdout = "0.1"
 
 [features]
 default = []
-api = []
