+++ examples/rust/Cargo.toml
@@ -5,4 +5,10 @@ authors = ["Jonny Rhea <jonny.rhea@consensys.net>"]
 edition = "2018"
 
 [dependencies]
-mothra = { path = "../../core", package = "mothra", features = ["api"]}
+mothra = { path = "../../core", package = "mothra"}
+tokio = "0.1.16"
+tokio-timer = "0.2.10"
+env_logger = "0.6.0"
+slog = { version = "^2.2.3" , features = ["max_level_trace"] }
+slog-term = "^2.4.0"
+slog-async = "^2.3.0"
